requirements.txt
flask
flask_sqlalchemy
flask_migrate
alembic
matplotlib
werkzeug
gunicorn
